Add Apple MPS support
MPS doesn't support float64, and rl_games hit it in two places:

* `RunningMeanStd` registers running_mean/var/count as float64 buffers,
  which crashes `.to('mps')`. The module is wrapped in
  `torch.jit.script`, so an `_apply` override is bypassed -- the dtype
  has to be decided at construction time. Pick float32 only on a Mac
  without CUDA where MPS is available; CUDA/CPU hosts keep float64 for
  the existing Welford precision.

* `Envpool.scores` and `Envpool.returned_scores` defaulted to numpy
  float64. The `DefaultAlgoObserver` then moves them to the training
  device, which throws on MPS. Allocate as `np.float32`.

Verified end-to-end with envpool 1.2.5 on Apple Silicon: Ant-v5
(continuous, MuJoCo) and Pong-v5 (discrete, Atari CNN) both train to
completion on `mps:0`; full Pong run hits score_to_win=20 at epoch
179/250 in ~3 minutes at ~7.5k total fps.

+++ rl_games/envs/envpool.py
@@ -62,8 +62,10 @@ class Envpool(IVecEnv):
 
         self.action_space = self.env.action_space
         self.ids = np.arange(0, num_actors)
-        self.scores = np.zeros(num_actors)
-        self.returned_scores = np.zeros(num_actors)
+        # float32 so downstream torch.from_numpy(...).to(device) works on MPS,
+        # which doesn't support float64.
+        self.scores = np.zeros(num_actors, dtype=np.float32)
+        self.returned_scores = np.zeros(num_actors, dtype=np.float32)
 
     def _set_scores(self, infos, dones):
         """Track raw (unclipped) episode rewards for Atari benchmarking.
